Update devcontainer and Makefile
This commit updates the devcontainer.json file to use the "make venv" command as the postCreateCommand. It also adds a new target "venv" to the Makefile, which creates a virtual environment and installs the required dependencies. Additionally, the requirements_dev.txt file is updated with new packages: PyPDF2, beautifulsoup4, and ipykernel.